# slivky/src/main/java/ru/urfu/slivky/web/dto/ShelterCreateRequest.java
package ru.urfu.slivky.web.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record ShelterCreateRequest(
        @NotBlank @Size(max = 255) String name,
        String description,
        String address,
        String phone,
        String email,
        String website
) {
}